ssp_exit: error messages and raise exception
This commit changes the ssp_exit function to accept a string as an error
message. If the status is a string, it is used as an error message and
the exit status is set to 1 (failure). The function now also accepts an
extra_message parameter, which is printed to stderr before exiting.

If the program is not running from the command line and the status
corresponds to an error, the function raises a RuntimeError exception.